+++ app/main.py
@@ -8,12 +8,13 @@ import requests
 from kivymd.app import MDApp
 from kivymd.uix.boxlayout import MDBoxLayout
 from kivymd.uix.navigationdrawer import MDNavigationDrawerMenu
-from kivymd.uix.menu import MDDropdownMenu
+#from kivymd.uix.menu import MDDropdownMenu
 from kivymd.uix.filemanager import MDFileManager
 from kivymd.uix.label import MDLabel
 from kivymd.uix.dialog import MDDialog
 from kivymd.uix.button import MDFlatButton, MDFloatingActionButton
 
+from kivy.uix.camera import Camera
 from kivy.clock import Clock
 from kivy.utils import platform
 from kivy.core.window import Window
@@ -22,6 +23,7 @@ from kivy.lang import Builder
 from kivy.uix.behaviors import ButtonBehavior
 from kivy.uix.image import Image
 from kivy.properties import StringProperty, NumericProperty, ObjectProperty, BooleanProperty
+from kivy.uix.widget import Widget
 
 from plyer import filechooser
 
@@ -29,13 +31,14 @@ from plyer import filechooser
 from services.faceAi import FaceAiSvc
 from screens.init_screen import ModelDownloder
 from screens.face_match import TempSpinWait, FaceMatchBox
+from screens.security import SecCamBox, SecCamBtn, NameInput, SecurityBox, SecAfterLogin, SecSingleFile
 from screens.setting import SettingsBox
 
 # IMPORTANT: Set this property for keyboard behavior
 Window.softinput_mode = "below_target"
 
 ## Global definitions
-__version__ = "0.0.1" # App version
+__version__ = "0.1.0" # App version
 
 # Determine the base path for your application's resources
 if getattr(sys, 'frozen', False):
@@ -46,9 +49,9 @@ else:
     base_path = os.path.dirname(os.path.abspath(__file__))
 kv_file_path = os.path.join(base_path, 'main_layout.kv')
 
-# imprt platform specific modules
+# import platform specific modules
 if platform == "android":
-    from jnius import autoclass, PythonJavaClass, java_method
+    from jnius import autoclass, cast, PythonJavaClass, java_method
     from android.permissions import check_permission, request_permissions, Permission
 
 ## -- kivy custom classes -- ##
@@ -87,8 +90,19 @@ class FaceAiApp(MDApp):
         self.is_onnx_running = False
         self.models_ok = False
         self.last_instance = None
+        self.wake_lock = None
+        self.sec_uix = None
+        self.camera = None
+        self.db_conn_ok = False
+        self.data_in_db = False
+        self.tmp_spinner = None
+        self.login_success = False
+        self.face_reg_error = False
+        self.sec_file_box = None
+        self.sec_file_current_path = None
         self.user_preferences = {
-            "fm_dont_again": False
+            "fm_dont_again": False,
+            "sec_dont_again": False
         }
 
     def build(self):
@@ -111,13 +125,16 @@ class FaceAiApp(MDApp):
             except Exception as e:
                 print(f"Could not check the android SDK version: {e}")
                 #self.show_toast_msg(f"Error checking SDK: {e}", is_error=True)
-            self.android_permissions = [] #Permission.CAMERA
+            self.android_permissions = [Permission.WAKE_LOCK]
+            self.total_permissions = [Permission.CAMERA]
             if sdk_version >= 33:  # Android 13+
                 self.android_permissions.append(Permission.READ_MEDIA_IMAGES)
             else:  # Android 9–12
                 self.android_permissions.extend([Permission.READ_EXTERNAL_STORAGE, Permission.WRITE_EXTERNAL_STORAGE])
-            request_permissions(self.android_permissions)
+            self.total_permissions.extend(self.android_permissions)
+            request_permissions(self.total_permissions)
             self.internal_storage = app_storage_path()
+            self.encrypt_key = None
             try:
                 Environment = autoclass("android.os.Environment")
                 self.external_storage = Environment.getExternalStorageDirectory().getAbsolutePath()
@@ -127,10 +144,13 @@ class FaceAiApp(MDApp):
         else:
             self.internal_storage = self.user_data_dir
             self.external_storage = os.path.expanduser("~")
+            self.encrypt_key = os.path.join(self.internal_storage, 'config', 'vault.key')
         # remaining start activities
         self.model_dir = os.path.join(self.internal_storage, 'model_files')
         self.op_dir = os.path.join(self.internal_storage, 'outputs')
         self.config_dir = os.path.join(self.internal_storage, 'config')
+        self.sec_file_dir = os.path.join(self.internal_storage, 'secure_files')
+        db_dir = os.path.join(self.internal_storage, 'databases')
         self.last_upload_path = None
         self.last_downlaod_path = None
         self.src_image_path = None
@@ -140,12 +160,15 @@ class FaceAiApp(MDApp):
         os.makedirs(self.model_dir, exist_ok=True)
         os.makedirs(self.op_dir, exist_ok=True)
         os.makedirs(self.config_dir, exist_ok=True)
+        os.makedirs(self.sec_file_dir, exist_ok=True)
+        os.makedirs(db_dir, exist_ok=True)
         print(f"Internal model files to be stored in: {self.model_dir}")
         self.config_path = os.path.join(self.config_dir, 'config.json')
         self.resp_path = os.path.join(self.config_dir, 'resp.json')
         self.usr_pref_path = os.path.join(self.config_dir, 'preferences.json')
         self.detect_model_path = os.path.join(self.model_dir, "faceai", "det_10g.onnx")
         self.recog_model_path = os.path.join(self.model_dir, "faceai","arc.onnx")
+        self.db_path = os.path.join(db_dir, "master_face.db")
         self.is_inp_file_mgr_open = False
         self.is_op_file_mgr_open = False
         # folder manager for downloading
@@ -160,12 +183,62 @@ class FaceAiApp(MDApp):
             with open(self.usr_pref_path, "r") as pf:
                 old_pref = json.load(pf)
             self.user_preferences["fm_dont_again"] = old_pref.get("fm_dont_again", False)
+            self.user_preferences["sec_dont_again"] = old_pref.get("sec_dont_again", False)
         else:
             with open(self.usr_pref_path, "w") as pf:
                 json.dump(self.user_preferences, pf)
+        # set encryption key for non android
+        if self.encrypt_key:
+            if not os.path.exists(self.encrypt_key):
+                from cryptography.fernet import Fernet
+                key = Fernet.generate_key()
+                with open(self.encrypt_key, "wb") as key_file:
+                    key_file.write(key)
+                print("Generated the encryption key")
         # check if model files are present
         Clock.schedule_once(lambda dt: self.model_existance_check())
         print("Init success...")
+
+    def acquire_wakelock(self):
+        if self.wake_lock:
+            return  # already acquired
+        try:
+            PythonActivity = autoclass("org.kivy.android.PythonActivity")
+            Context = autoclass("android.content.Context")
+            activity = PythonActivity.mActivity
+            PowerManager = autoclass("android.os.PowerManager")
+            power_manager = cast(PowerManager, activity.getSystemService(Context.POWER_SERVICE))
+            # Create wakelock (use PowerManager.FULL_WAKE_LOCK for full wakelock)
+            self.wake_lock = power_manager.newWakeLock(
+                PowerManager.FULL_WAKE_LOCK, "MyApp::WakeLockTag"
+            )
+            self.wake_lock.acquire()
+            print("WakeLock acquired")
+        except Exception as e:
+            print(f"Wake lock aquire error: {e}")
+
+    def release_wakelock(self):
+        if self.wake_lock and self.wake_lock.isHeld():
+            self.wake_lock.release()
+            self.wake_lock = None
+            print("WakeLock released")
+
+    def save_config_file(self, configData:dict, filepath:str):
+        with open(filepath, "w") as pf:
+            json.dump(configData, pf)
+
+    def set_usr_pref(self, instance=None):
+        self.txt_dialog_closer()
+        saveFlag = False
+        if self.root.ids.screen_manager.current == "faceFindScr":
+            self.user_preferences["fm_dont_again"] = True
+            saveFlag = True
+        elif self.root.ids.screen_manager.current == "securityScr":
+            self.user_preferences["sec_dont_again"] = True
+            saveFlag = True
+        # save the file
+        if saveFlag:
+            Clock.schedule_once(lambda dt: self.save_config_file(self.user_preferences, self.usr_pref_path))
 
     def check_request_android_permission(self):
         if platform == "android":
@@ -177,6 +250,20 @@ class FaceAiApp(MDApp):
                     break
             if not permission_flag:
                 request_permissions(self.android_permissions, self.permission_callback)
+            return permission_flag
+        else:
+            return True
+
+    def check_request_total_permission(self):
+        if platform == "android":
+            permission_flag = True
+            for permission in self.total_permissions:
+                tmp_flag = check_permission(permission)
+                if not tmp_flag:
+                    permission_flag = False
+                    break
+            if not permission_flag:
+                request_permissions(self.total_permissions, self.permission_callback)
             return permission_flag
         else:
             return True
@@ -270,7 +357,7 @@ class FaceAiApp(MDApp):
                 on_release=self.txt_dialog_closer
             ),
             MDFlatButton(
-                text="Ok",
+                text="Download",
                 theme_text_color="Custom",
                 text_color="green",
                 on_release=self.start_model_download
@@ -278,7 +365,7 @@ class FaceAiApp(MDApp):
         ]
         self.show_text_dialog(
             "Download model files",
-            "You need to downlaod the models first.",
+            "You need to downlaod the models first (~140MB).",
             buttons
         )
 
@@ -290,6 +377,8 @@ class FaceAiApp(MDApp):
             os.remove(filepath)
             self.show_toast_msg("Model files have been downloaded successfully.")
             self.is_downloading = False
+            if platform == "android":
+                Clock.schedule_once(lambda dt: self.release_wakelock())
         except Exception as e:
             print(f"Unzip error: {e}")
 
@@ -342,6 +431,8 @@ class FaceAiApp(MDApp):
         if self.is_downloading:
             self.show_toast_msg("Please wait for the download to finish", is_error=True)
             return
+        if platform == "android":
+            Clock.schedule_once(lambda dt: self.acquire_wakelock())
         download_path = os.path.join(self.model_dir, "faceai.tar.gz")
         model_url = "https://github.com/daslearning-org/face-ai/releases/download/vOnnxModels/faceai.tar.gz"
         self.download_model_file(model_url, download_path)
@@ -377,6 +468,29 @@ class FaceAiApp(MDApp):
 
     def goto_face_matcher(self, instance=None):
         self.root.ids.screen_manager.current = "faceFindScr"
+
+    def on_face_matcher_entry(self, instance=None):
+        if self.user_preferences["fm_dont_again"]:
+            return
+        buttons = [
+            MDFlatButton(
+                text="Don't Show Again",
+                theme_text_color="Custom",
+                text_color='orange',
+                on_release=self.set_usr_pref
+            ),
+            MDFlatButton(
+                text="Ok",
+                theme_text_color="Custom",
+                text_color="green",
+                on_release=self.txt_dialog_closer
+            ),
+        ]
+        self.show_text_dialog(
+            "Instructions",
+            "Upload an image with Single Face as Source image & any image with one or more Faces as Target image. Once the face tagged images are generated, you can tap on that to Downlaod or Delete.",
+            buttons
+        )
 
     def open_img_file_manager(self, instance=None):
         permissions_ok = self.check_request_android_permission()
@@ -460,6 +574,10 @@ class FaceAiApp(MDApp):
         if not self.src_image_path or not self.trgt_image_path:
             self.show_toast_msg("Please upload both the source & target image", is_error=True)
             return
+        src_box = self.root.ids.face_match_scr.ids.fm_gen_src_box
+        trgt_box = self.root.ids.face_match_scr.ids.fm_gen_trgt_box
+        src_box.clear_widgets()
+        trgt_box.clear_widgets()
         onnx_thread = Thread(
             target=self.face_ai.face_match_group,
             kwargs={
@@ -471,6 +589,10 @@ class FaceAiApp(MDApp):
         )
         onnx_thread.start()
         self.is_onnx_running = True
+        spinner1 = TempSpinWait()
+        spinner2 = TempSpinWait()
+        src_box.add_widget(spinner1)
+        trgt_box.add_widget(spinner2)
 
     def face_match_callback(self, resp):
         self.is_onnx_running = False
@@ -501,7 +623,7 @@ class FaceAiApp(MDApp):
             label = MDLabel(
                 text=msg,
                 halign="center",
-                #valign="top",
+                valign="center",
                 markup=True
             )
             src_box.add_widget(label)
@@ -522,16 +644,16 @@ class FaceAiApp(MDApp):
                         on_release=self.txt_dialog_closer
                     ),
                     MDFlatButton(
-                        text="Download",
-                        theme_text_color="Custom",
-                        text_color="green",
-                        on_release=self.download_from_app_local
-                    ),
-                    MDFlatButton(
                         text="DELETE",
                         theme_text_color="Custom",
                         text_color="red",
                         on_release=self.delete_selected_file
+                    ),
+                    MDFlatButton(
+                        text="Download",
+                        theme_text_color="Custom",
+                        text_color="green",
+                        on_release=self.download_from_app_local
                     ),
                 ],
             )
@@ -554,14 +676,30 @@ class FaceAiApp(MDApp):
         dir_name = os.path.dirname(path)
         self.last_downlaod_path = dir_name
         chosen_path = os.path.join(path, filename) # destination path
-        import shutil
         try:
-            shutil.copyfile(self.download_file_path, chosen_path)
+            if (self.root.ids.screen_manager.current in ("faceFindScr") or 
+                platform == "android"
+            ):
+                import shutil
+                shutil.copyfile(self.download_file_path, chosen_path)
+                self.delete_file_path = str(self.download_file_path)
+                self.download_file_path = None
+                self.delete_file_popup()
+            elif (self.root.ids.screen_manager.current in ("securityScr") and 
+                  platform != "android"
+                ):
+                # decrypt the data
+                from cryptography.fernet import Fernet
+                with open(self.encrypt_key, "rb") as key_file:
+                    key = key_file.read()
+                fernet = Fernet(key)
+                with open(self.download_file_path, "rb") as enc_file:
+                    encrypted_data = enc_file.read()
+                decrypted_data = fernet.decrypt(encrypted_data)
+                with open(chosen_path, "wb") as file:
+                    file.write(decrypted_data)
             print(f"File successfully download to: {chosen_path}")
             self.show_toast_msg(f"File download to: {chosen_path}")
-            self.delete_file_path = str(self.download_file_path)
-            self.download_file_path = None
-            self.delete_file_popup()
         except Exception as e:
             print(f"Error saving file: {e}")
             self.show_toast_msg(f"Error saving file: {e}", is_error=True)
@@ -600,10 +738,9 @@ class FaceAiApp(MDApp):
                 self.show_toast_msg(f"Deleted: {self.delete_file_path}")
                 print(f"Deleted: {self.delete_file_path}")
                 self.delete_file_path = None
-                self.download_file_path = None
                 if self.last_instance and self.last_instance.parent:
-                    self.last_instance.parent.remove_widget(self.last_instance)
-                    self.last_instance = None
+                    Clock.schedule_once(lambda dt: self.last_instance.parent.remove_widget(self.last_instance))
+                    #self.last_instance = None
             except Exception as e:
                 print(f"Error while deleting {self.delete_file_path} because: {e}")
                 self.show_toast_msg(f"Error while deleting {self.delete_file_path} because: {e}", is_error=True)
@@ -621,6 +758,312 @@ class FaceAiApp(MDApp):
         upload_src_box.clear_widgets()
         upload_trgt_box.clear_widgets()
 
+    # Security section starts here
+    def on_security_enter(self):
+        permissions_ok = self.check_request_total_permission()
+        if not permissions_ok:
+            return
+        if not self.face_ai:
+            self.show_toast_msg("Please start the services first.", is_error=True, duration=2)
+            self.root.ids.screen_manager.current = "initScreen"
+            return
+        self.on_security_leave() # clear existing things
+        self.sec_uix = self.root.ids.security_box.ids.sec_elements_box
+        self.db_conn_ok = self.face_ai.start_db_session(self.db_path)
+        if self.db_conn_ok:
+            self.tmp_spinner = TempSpinWait()
+            self.tmp_spinner.text = "Checking database, please wait..."
+            self.sec_uix.add_widget(self.tmp_spinner)
+            height_widget = Widget(
+                size_hint_y = 0.5
+            )
+            self.sec_uix.add_widget(height_widget)
+            Clock.schedule_once(lambda dt: self.face_ai.check_if_data_exist(self._init_security_callback))
+
+    def android_front_cam(self):
+        from services.androidSvc import AndroidSvc
+        cam_svc = AndroidSvc()
+        cam_map = cam_svc.get_camera_details()
+        cam_id = cam_map.get("front", None)
+        return cam_id
+
+    def add_camera(self, parent_element):
+        #self.sec_uix = self.root.ids.security_box.ids.camera_box
+        if self.sec_uix:
+            self.sec_uix.clear_widgets()
+        self.sec_uix = parent_element
+        cam_indx = 0
+        if platform == "android":
+            cam_id = self.android_front_cam()
+            if cam_id:
+                cam_indx = int(cam_id)
+            resolution = (960, 720) # will fallback to 480 if fails again
+        else:
+            resolution = (640, 480)
+            if not self.camera:
+                import cv2
+                available_cameras = []
+                for i in range(2):
+                    cap = cv2.VideoCapture(i)
+                    if cap.isOpened():
+                        print(f"Camera found at index: {i}")
+                        available_cameras.append(i)
+                        cap.release()
+                if len(available_cameras) >= 1:
+                    cam_indx = available_cameras[0]
+                else:
+                    self.show_toast_msg(f"No camera found on {platform}!", is_error=True)
+                    return
+        try:
+            if not self.camera:
+                self.camera = Camera(
+                    index = cam_indx,
+                    resolution = resolution,
+                    fit_mode = "contain",
+                    play = True
+                )
+            else:
+                self.camera.play = True
+            self.sec_uix.add_widget(self.camera)
+        except Exception as e:
+            print(f"Error setting up the camera: {e}")
+            self.show_toast_msg(f"Error setting up the camera: {e}", is_error=True)
+
+    def add_name_input(self, parent_element):
+        name_inp_elem = NameInput()
+        parent_element.add_widget(name_inp_elem)
+
+    def add_login_btn(self, parent_element):
+        login_btn = SecCamBtn()
+        height_widget = Widget(
+            size_hint_y = 0.5
+        )
+        parent_element.add_widget(login_btn)
+        parent_element.add_widget(height_widget)
+
+    def _init_security_callback(self, resp): # should not be called directly
+        if self.login_success:
+            self.sec_login_ok()
+            return
+        if self.sec_uix:
+            self.sec_uix.clear_widgets()
+        if not resp: # no data found 
+            self.add_camera(self.root.ids.security_box.ids.sec_elements_box)
+            self.add_name_input(self.root.ids.security_box.ids.sec_elements_box)
+        else: # data found, need login
+            self.data_in_db = True
+            self.add_camera(self.root.ids.security_box.ids.sec_elements_box)
+            self.add_login_btn(self.root.ids.security_box.ids.sec_elements_box)
+
+    def add_new_face(self, instance=None):
+        self.on_security_leave()
+        self.add_camera(self.root.ids.security_box.ids.sec_elements_box)
+        self.add_name_input(self.root.ids.security_box.ids.sec_elements_box)
+
+    def get_sec_file_list(self, path:str):
+        files = []
+        for filename in os.listdir(path):
+            files.append(filename)
+        return files
+
+    def refresh_sec_file_list(self, instance=None):
+        if self.sec_file_box:
+            secMdList = self.sec_file_box.ids.sec_file_list
+            secMdList.clear_widgets()
+            sec_files = self.get_sec_file_list(self.sec_file_dir)
+            for file in sec_files:
+                fileElem = SecSingleFile()
+                fileElem.filename = str(file)
+                secMdList.add_widget(fileElem)
+
+    def after_login_instructions(self):
+        if self.user_preferences["sec_dont_again"]:
+            return
+        else:
+            buttons = [
+                MDFlatButton(
+                    text="Don't Show Again",
+                    theme_text_color="Custom",
+                    text_color='orange',
+                    on_release=self.set_usr_pref
+                ),
+                MDFlatButton(
+                    text="Ok",
+                    theme_text_color="Custom",
+                    text_color="green",
+                    on_release=self.txt_dialog_closer
+                ),
+            ]
+            self.show_text_dialog(
+                "Instructions",
+                "You can upload your secret files here. You can download them later. You can add a new face for login.",
+                buttons
+            )
+
+    def sec_login_ok(self, msg:str="Login"):
+        if self.camera:
+            self.camera.play = False
+            #self.camera._camera.stop()
+            #self.camera = None
+        if self.sec_uix:
+            self.sec_uix.clear_widgets()
+        self.sec_file_box = SecAfterLogin()
+        secMdList = self.sec_file_box.ids.sec_file_list
+        sec_files = self.get_sec_file_list(self.sec_file_dir)
+        for file in sec_files:
+            fileElem = SecSingleFile()
+            fileElem.filename = str(file)
+            secMdList.add_widget(fileElem)
+        self.sec_uix.add_widget(self.sec_file_box)
+        self.login_success = True
+        self.after_login_instructions()
+        print(f"{msg} is successful!")
+
+    def sec_face_login_save(self, name:str, img, newFace=False):
+        if not self.data_in_db or newFace:
+            matched_name = None
+            if len(name) < 3:
+                self.show_toast_msg("Please enter a proper name", True)
+                return
+            if self.data_in_db:
+                matched_name = self.face_ai.face_verify(img)
+            if matched_name is None:
+                stat = self.face_ai.save_faces_masterdb(name, img)
+                if stat:
+                    self.data_in_db = True
+                    self.sec_login_ok("SignUp")
+                    Clock.schedule_once(lambda dt: self.show_toast_msg(f"{name}'s face has been added in database."))
+                else:
+                    self.show_toast_msg("Face is not saved, please try again", True)
+                    self.face_reg_error = True
+            else:
+                msg = f"This face is already saved for: {str(matched_name[0])}"
+                self.sec_login_ok()
+                Clock.schedule_once(lambda dt: self.show_toast_msg(msg))
+        else:
+            print("There is existing face(s)")
+            matched_name = self.face_ai.face_verify(img)
+            if matched_name is None:
+                self.show_toast_msg("Face is not matched", True)
+            else:
+                msg = f"Logged in as: {str(matched_name[0])}"
+                self.sec_login_ok()
+                Clock.schedule_once(lambda dt: self.show_toast_msg(msg))
+
+    def sec_capture_frame(self, instance=None, newFace=False):
+        if not self.camera or not self.camera.texture:
+            self.show_toast_msg("Camera is Not OK", True, 2)
+            return
+        if instance:
+            name_txt = str(instance.text)
+            name_txt = name_txt.strip()
+        else:
+            name_txt = ""
+        try:
+            import cv2
+            import numpy as np
+            texture = self.camera.texture
+            pixels = texture.pixels
+            width, height = texture.size
+            arr = np.frombuffer(pixels, dtype=np.uint8).reshape((height, width, 4))
+            if platform == 'android':
+                arr = np.flipud(arr)  # Flip up down in android
+            img = cv2.cvtColor(arr, cv2.COLOR_RGBA2BGR)
+            #print(img)
+            self.sec_face_login_save(name_txt, img, newFace)
+        except Exception as e:
+            print(f"Error processing frame: {e}")
+
+    def vault_file_saver(self, file_path:str):
+        filename = os.path.basename(file_path)
+        target_file = os.path.join(self.sec_file_dir, filename)
+        try:
+            if platform == "android":
+                # encryption not required on android as the internal path is secure
+                import shutil
+                shutil.copyfile(file_path, target_file)
+            else:
+                # store encrypted data
+                from cryptography.fernet import Fernet
+                with open(self.encrypt_key, "rb") as key_file:
+                    key = key_file.read()
+                fernet = Fernet(key)
+                with open(file_path, "rb") as file:
+                    original_data = file.read()
+                encrypted_data = fernet.encrypt(original_data)
+                with open(target_file, "wb") as tfile:
+                    tfile.write(encrypted_data)
+            print(f"File {filename} successfully uploaded to vault")
+            self.refresh_sec_file_list()
+            self.show_toast_msg(f"File {filename} successfully uploaded to vault")
+        except Exception as e:
+            print(f"Error saving file: {e}")
+            self.show_toast_msg(f"Error saving file: {e}", is_error=True)
+
+    def save_in_vault(self, selection):
+        self.is_inp_file_mgr_open = False
+        if selection:
+            file_path = str(selection[0])
+            self.last_upload_path = os.path.dirname(file_path)
+            Clock.schedule_once(lambda dt: self.vault_file_saver(file_path))
+
+    def upload_to_vault(self, instance=None):
+        permissions_ok = self.check_request_android_permission()
+        if not permissions_ok:
+            return
+        try:
+            if not self.last_upload_path:
+                self.last_upload_path = self.external_storage
+            filechooser.open_file(
+                on_selection = self.save_in_vault,
+                path = self.last_upload_path,
+                multiple = False,
+                filters = [["*.JPG", "*.jpg", "*.png", "*.jpeg", "*.webp"], "*"],
+                preview = True,
+            )
+            self.is_inp_file_mgr_open = True
+        except Exception as e:
+            self.show_toast_msg(f"Error: {e}", is_error=True)
+
+    def download_sec_file(self, instance):
+        filename = str(instance.filename)
+        full_path = os.path.join(self.sec_file_dir, filename)
+        self.download_file_path = str(full_path)
+        self.download_from_app_local()
+
+    def popup_sec_file_delete(self, instance):
+        filename = str(instance.filename)
+        self.last_instance = instance
+        self.delete_file_path = os.path.join(self.sec_file_dir, filename)
+        self.show_text_dialog(
+            title="Delete the file?",
+            text=f"Do you want to delete: {filename} ?",
+            buttons=[
+                MDFlatButton(
+                    text="Cancel",
+                    theme_text_color="Custom",
+                    text_color="blue",
+                    on_release=self.txt_dialog_closer
+                ),
+                MDFlatButton(
+                    text="DELETE",
+                    theme_text_color="Custom",
+                    text_color="red",
+                    on_release=self.delete_selected_file
+                ),
+            ],
+        )
+
+    def on_security_leave(self):
+        if self.camera:
+            self.camera.play = False
+            #self.camera._camera.stop()
+            #self.camera = None
+            print("Camera has been stopped")
+        if self.sec_uix:
+            self.sec_uix.clear_widgets()
+
+    # Settings section start here
     def show_all_delete_alert(self):
         op_img_count = 0
         for filename in os.listdir(self.op_dir):
@@ -688,12 +1131,26 @@ class FaceAiApp(MDApp):
             buttons
         )
 
+    # Handling the device events (mostly on Android)
     def on_pause(self):
+        self.login_success = False
         return True
 
     def events(self, instance, keyboard, keycode, text, modifiers):
         """Handle mobile device button presses (e.g., Android back button)."""
-        #if keyboard in (1001, 27):
+        if keyboard in (1001, 27):
+            # control file manager with back btn on android
+            if self.is_op_file_mgr_open:
+                if self.op_file_manager.current_path == self.external_storage:
+                    self.show_toast_msg(f"Closing file manager from main storage")
+                    self.op_file_exit_manager()
+                else:
+                    self.op_file_manager.back() # go one dir back
+                # stop app from exiting
+                return True
+            if self.face_reg_error:
+                self.on_security_enter()
+        # exits from app
         return False
 
 ## -- run the app -- ##