+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 ## Changes
 
+### 0.1.0
+- Adding a `Secure Vault` feature to store secret files.
+- Pop-Up with a quick tip for the chosen feature.
+- Changed the documentation link.
+- Some UI enhancements.
+
 ### 0.0.1
-- The initial version (WIP)
+- The initial version